Data Acquistion: use full packet in callback, not just data buffer
To make it easier to work with the data from our callback system, we
send back the bytes buffer for the full packet (including header) in
the callback.  This matches the output of downloading the data file,
and makes it easier to write downstream consumers (they now just read
the file as if it were a bunch of packets).  Current experimental data
readout code is doing minimal parsing, partly because the CRCs are
wonky, but mosly because it's a heck of a lot faster when you're
churning thorugh the tens of thousands of packets of data in a local
file.